Acme Bank v1.0.0-dev Backoffice.Router.Helpers

Module with named helpers generated from Backoffice.Router.

Functions

admin_association_path(conn_or_endpoint, action, resource, id, association_name)

admin_association_path(conn_or_endpoint, action, resource, id, association_name, params)

admin_association_url(conn_or_endpoint, action, resource, id, association_name)

admin_association_url(conn_or_endpoint, action, resource, id, association_name, params)

admin_path(conn_or_endpoint, action)

admin_path(conn_or_endpoint, action, params)

admin_path(conn_or_endpoint, action, id, params)

admin_resource_path(conn_or_endpoint, action, resource)

admin_resource_path(conn_or_endpoint, action, resource, params)

admin_resource_path(conn_or_endpoint, action, resource, id, params)

admin_resource_path(conn_or_endpoint, action, resource, id, action, params)

admin_resource_url(conn_or_endpoint, action, resource)

admin_resource_url(conn_or_endpoint, action, resource, params)

admin_resource_url(conn_or_endpoint, action, resource, id, params)

admin_resource_url(conn_or_endpoint, action, resource, id, action, params)

admin_url(conn_or_endpoint, action)

admin_url(conn_or_endpoint, action, params)

admin_url(conn_or_endpoint, action, id, params)

path(data, path)

Generates the path information including any necessary prefix.

static_path(conn, path)

Generates path to a static asset given its file path.

static_url(conn, path)

Generates url to a static asset given its file path.

url(data)

Generates the connection/endpoint base URL without any path information.