Acme Bank v1.0.0-dev BankWeb.Router.Helpers

Module with named helpers generated from BankWeb.Router.

Functions

account_path(conn_or_endpoint, action)

account_path(conn_or_endpoint, action, params)

account_url(conn_or_endpoint, action)

account_url(conn_or_endpoint, action, params)

page_path(conn_or_endpoint, action)

page_path(conn_or_endpoint, action, params)

page_url(conn_or_endpoint, action)

page_url(conn_or_endpoint, action, params)

path(data, path)

Generates the path information including any necessary prefix.

session_path(conn_or_endpoint, action)

session_path(conn_or_endpoint, action, params)

session_path(conn_or_endpoint, action, username, params)

session_url(conn_or_endpoint, action)

session_url(conn_or_endpoint, action, params)

session_url(conn_or_endpoint, action, username, params)

static_path(conn, path)

Generates path to a static asset given its file path.

static_url(conn, path)

Generates url to a static asset given its file path.

transfer_path(conn_or_endpoint, action)

transfer_path(conn_or_endpoint, action, params)

transfer_url(conn_or_endpoint, action)

transfer_url(conn_or_endpoint, action, params)

url(data)

Generates the connection/endpoint base URL without any path information.